requirements.txt
streamlit
torch
ultralytics
Pillow
